requirements.txt
greenlet==3.2.4
mysql-connector-python==9.4.0
numpy==2.3.2
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
tzdata==2025.2
